backend/pyproject.toml
[project]
name = "docforge"
version = "0.1.0"
description = "Template-driven document generation engine"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn[standard]>=0.30.0",
    "python-docx>=1.1.0",
    "lxml>=5.0.0",
    "pandas>=2.2.0",
    "openpyxl>=3.1.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy>=2.0.30",
    "python-multipart>=0.0.9",
    "click>=8.1.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
docforge = "cli:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"api/*.py" = ["B008"]  # Allow FastAPI's File(), Depends() in defaults

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.setuptools.packages.find]
include = ["api*", "core*", "renderers*", "extractors*", "db*", "utils*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
